@@ -3,6 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/react-loadable": "^5.5.3",
+    "react-loadable": "^5.5.0",
+    "react-swipeable-views": "^0.13.9",
+    "react-webcam": "^5.0.2",
     "serve": "^11.2.0"
   },
   "scripts": {
@@ -11,6 +15,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "start": "serve -s build",
+    "win-dev-https": "($env:HTTPS = \"true\") -and (npm run dev)",
     "heroku-postbuild": "npm run build"
   },
   "eslintConfig": {
@@ -29,42 +34,47 @@
     ]
   },
   "devDependencies": {
-    "lodash": "^4.17.15",
-    "@feathersjs/authentication-client": "^4.3.11",
-    "@feathersjs/client": "^4.3.10",
-    "@feathersjs/feathers": "^4.3.11",
-    "@feathersjs/socketio-client": "^4.3.11",
-    "@material-ui/core": "^4.8.3",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.34",
+    "@feathersjs/authentication-client": "^4.5.3",
+    "@feathersjs/client": "^4.5.3",
+    "@feathersjs/feathers": "^4.5.3",
+    "@feathersjs/socketio-client": "^4.5.3",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@fortawesome/react-fontawesome": "^0.1.9",
+    "@material-ui/core": "^4.9.12",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.51",
     "@types/is-uuid": "^1.0.0",
-    "@types/jest": "24.0.19",
-    "@types/lodash": "^4.14.149",
+    "@types/jest": "25.2.1",
+    "@types/lodash": "^4.14.150",
     "@types/node": "12.7.12",
-    "@types/react": "16.9.7",
-    "@types/react-dom": "16.9.2",
-    "@types/react-redux": "^7.1.5",
-    "@types/react-router": "^5.1.3",
-    "@types/react-router-dom": "^5.1.3",
+    "@types/react": "16.9.34",
+    "@types/react-dom": "16.9.7",
+    "@types/react-fontawesome": "^1.6.4",
+    "@types/react-redux": "^7.1.7",
+    "@types/react-router": "^5.1.7",
+    "@types/react-router-dom": "^5.1.5",
+    "@types/react-swipeable-views": "^0.13.0",
     "@types/seedrandom": "^2.4.28",
     "@types/socket.io-client": "^1.4.32",
-    "@types/uuid": "^3.4.6",
+    "@types/uuid": "^3.4.9",
     "is-uuid": "^1.0.2",
-    "material-table": "^1.54.1",
-    "node-sass": "^4.13.0",
-    "notistack": "^0.9.6",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
-    "react-redux": "^7.1.3",
+    "lodash": "^4.17.15",
+    "material-table": "^1.57.2",
+    "node-sass": "^4.14.0",
+    "notistack": "^0.9.9",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.0",
-    "redux": "^4.0.4",
+    "react-scripts": "^3.4.1",
+    "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "seedrandom": "^3.0.5",
     "socket.io-client": "^2.3.0",
-    "systeminformation": "^4.18.0",
-    "typescript": "3.6.4",
-    "uuid": "^3.3.3",
+    "systeminformation": "^4.23.8",
+    "typescript": "^3.8.3",
+    "uuid": "^3.4.0",
     "validate.js": "^0.13.1"
   }
 }
