@@ -1,27 +1,15 @@
-import * as lodash from 'lodash';
-import React from 'react';
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { ThunkDispatch } from 'redux-thunk';
 import {
   Typography,
-  Select,
   MenuItem,
   Button,
   TextField,
-  InputLabel,
-  FormControl,
-  Input,
+  Grid,
   Chip,
-  Checkbox,
-  ListItemText
 } from '@material-ui/core';
-import {
-  withStyles,
-  WithStyles,
-  createStyles,
-  StyleRules,
-  Theme
-} from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import Rating from '@material-ui/lab/Rating';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 
@@ -34,46 +22,40 @@ import { QuestionTypesState } from '../../../store/questionTypes/types';
 import { QuestionType } from '../../../api/classes/questionType.class';
 import { GameType } from '../../../api/classes/gameType.class';
 import { Helper } from '../../../helper';
+import { DangerButton } from '../utils/dangerButton.component';
+import { OutlinedDiv } from '../utils/outlinedDiv.component';
 
-const styles = (theme: Theme): StyleRules =>
-  createStyles({
-    form: {
-      display: 'flex',
-      flexDirection: 'column',
-      alignContent: 'center'
+const useStyles = makeStyles((theme) => ({
+  hotLevelRating: {
+    color: '#FD6C9E',
+  },
+  questionTypeContainer: {
+    padding: 2,
+  },
+  questionTypeChip: {
+    width: 'calc(50% - 10px)',
+    margin: 5,
+    marginBottom: 3,
+    '&:nth-child(-n+2)': {
+      marginTop: 7,
     },
-    formControl: {
-      paddingBottom: theme.spacing(1)
+    '&:nth-last-child(-n+2)': {
+      marginBottom: 5,
     },
-    ratingFormControl: {
-      padding: theme.spacing(1),
-      flexDirection: 'row'
-    },
-    ratingComponent: {
-      marginLeft: theme.spacing(2)
-    },
-    hotLevelRating: {
-      color: '#FD6C9E'
-    },
-    questionTypeSelected: {
-      fontWeight: theme.typography.fontWeightMedium
-    },
-    chips: {
-      display: 'flex',
-      flexWrap: 'wrap'
-    },
-    chip: {
-      margin: '2px',
-      minWidth: `calc(50% - 4px)`,
-      zIndex: 1200
-    }
-  });
+  },
+  unselectable: {
+    userSelect: 'none',
+  },
+}));
 
 interface OwnProps {
   game: Game;
-  onSubmit: (game: Game) => void;
-  buttonText?: string;
+  editable: boolean;
   disabled?: boolean;
+  acceptButtonText?: string;
+  deleteButtonText?: string;
+  onSubmit?: (game: Game) => void;
+  onDelete?: (gameId: string) => void;
 }
 
 interface DispatchProps {
@@ -86,330 +68,243 @@ interface StateProps {
   questionTypesState: QuestionTypesState;
 }
 
-type Props = StateProps & OwnProps & DispatchProps & WithStyles<typeof styles>;
+type Props = StateProps & OwnProps & DispatchProps;
 
-interface ComponentState {
-  game: Game;
-  error: string;
-}
+const GameFormComponent: React.FunctionComponent<Props> = (props: Props) => {
+  const classes = useStyles();
+  const allGameTypes = props.gameTypesState.gameTypes!;
+  const allQuestionsTypes = props.questionTypesState.questionTypes!;
 
-class GameFormComponent extends React.Component<Props, ComponentState> {
-  public static defaultProps = {
-    buttonText: 'Accept',
-    disabled: false
-  };
+  const [name, setName] = useState(props.game.name);
+  const [type, setType] = useState(
+    props.game.type.name !== '' ? props.game.type : allGameTypes[0]
+  );
+  const [maxDifficulty, setMaxDifficulty] = useState(props.game.maxDifficulty);
+  const [maxHotLevel, setMaxHotLevel] = useState(props.game.maxHotLevel);
+  const [nbTurns, setNbTurns] = useState(props.game.nbTurns);
+  const [questionTypes, setQuestionTypes] = useState(props.game.questionTypes);
 
-  /**
-   *
-   */
-  constructor(props: Props) {
-    super(props);
-
-    if (props.game.questionTypes.length === 0)
-      props.game.questionTypes = [...props.questionTypesState.questionTypes!];
-    if (GameType.CompareObjects(props.game.type, new GameType()))
-      props.game.type = props.gameTypesState.gameTypes![0];
-
-    this.state = {
-      game: props.game,
-      error: ''
-    };
-
-    this.isDenied = this.isDenied.bind(this);
-  }
-
-  private static prevPropsGame: Game;
-  static getDerivedStateFromProps(nextProps: Props, prevState: ComponentState) {
-    let nextPropsGame = nextProps.game;
-    if (!Game.CompareObjects(nextPropsGame, GameFormComponent.prevPropsGame)) {
-      GameFormComponent.prevPropsGame = lodash.cloneDeep(nextPropsGame);
-      return {
-        game: lodash.cloneDeep(nextPropsGame)
-      };
-    }
-
-    return null;
-  }
-
-  handleNameChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    this.setState({
-      game: Helper.clone(this.state.game, { name: e.target.value })
-    });
-  };
-
-  handleMaxDifficultyChange = (e: any, value: number) => {
-    this.setState({
-      game: Helper.clone(this.state.game, { maxDifficulty: value })
-    });
-  };
-
-  handleMaxHotLevelChange = (e: any, value: number) => {
-    this.setState({
-      game: Helper.clone(this.state.game, { maxHotLevel: value })
-    });
-  };
-
-  handleNbTurnsChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    this.setState({
-      game: Helper.clone(this.state.game, { nbTurns: parseInt(e.target.value) })
-    });
-  };
-
-  handleGameTypeChange = (e: React.ChangeEvent<{ value: unknown }>) => {
-    const gameTypes = this.props.gameTypesState.gameTypes;
-    if (gameTypes) {
-      this.setState({
-        game: Helper.clone(this.state.game, {
-          type: gameTypes.find(
-            t => t.id === parseInt(e.target.value as string)
-          )!
-        })
-      });
-    } else {
-      console.error('Game types not loaded yet !', 'How did you get here ?');
-    }
-  };
-
-  handleQuestionTypeChange = (e: React.ChangeEvent<{ value: any }>) => {
-    const questionTypes = this.props.questionTypesState.questionTypes;
-
-    if (questionTypes) {
-      let selectedTypes: QuestionType[] = e.target.value.filter(
-        (v: any) => v instanceof QuestionType
-      );
-      const clickedId: string = e.target.value.filter(
-        (v: any) => typeof v === 'string'
-      )[0];
-
-      const indexToRemove = selectedTypes.findIndex(t => t.id === clickedId);
-      if (indexToRemove === -1)
-        selectedTypes.push(questionTypes.find(t => t.id === clickedId)!);
-      else selectedTypes.splice(indexToRemove, 1);
-
-      this.setState({
-        game: Helper.clone(this.state.game, { questionTypes: selectedTypes })
-      });
-    } else {
-      console.error(
-        'Question types not loaded yet !',
-        'How did you get here ?'
-      );
-    }
-  };
-
-  removeQuestionType = (typeId: string) => {
-    this.setState({
-      game: Helper.clone(this.state.game, {
-        questionTypes: lodash.reject(
-          this.state.game.questionTypes,
-          t => t.id === typeId
-        )
-      })
-    });
-  };
-
-  handleFormSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    if (!this.isDenied()) {
-      this.props.onSubmit(this.state.game);
-    }
-  };
-
-  isDenied(): boolean {
-    const {
-      maxDifficulty,
-      maxHotLevel,
-      name,
-      nbTurns,
-      type,
-      questionTypes
-    } = this.state.game;
-    const gameTypes = this.props.gameTypesState.gameTypes;
-    const gameLoading = this.props.gamesState.loading;
-    let e: string = '';
-
-    if (gameLoading) e = 'Game infos are loading...';
-    if (!gameTypes!.find(t => GameType.CompareObjects(t, type)))
-      e = 'Incorrect type';
-    if (maxDifficulty === 0) e = 'Incorrect maximum difficulty';
-    if (maxHotLevel === 0) e = 'Incorrect maximum hot level';
-    if (name.length === 0) e = 'Incorrect name';
-    if (isNaN(nbTurns) || nbTurns <= 0) e = 'Incorrect nb turns';
-    if (questionTypes.length === 0) e = 'Incorrect question types';
-
-    if (e) console.log(e);
-
-    return e !== '';
-  }
-
-  isQuestionTypeSelected(questionType: QuestionType) {
-    return this.state.game.questionTypes.includes(questionType);
-  }
-
-  render() {
-    const classes = this.props.classes;
-    const {
-      maxDifficulty,
-      maxHotLevel,
-      name,
-      nbTurns,
-      type,
-      questionTypes
-    } = this.state.game;
-
-    const allGameTypes = this.props.gameTypesState.gameTypes;
-    const allQuestionTypes = this.props.questionTypesState.questionTypes;
+  const isDenied = (): boolean => {
+    const gameLoading = props.gamesState.loading;
 
     return (
-      <form className={classes.form} onSubmit={this.handleFormSubmit}>
-        <FormControl className={classes.formControl}>
-          <TextField
-            style={{ textAlign: 'left' }}
-            select
-            label="Game Type"
-            id="type-select"
-            value={type.id}
-            fullWidth
-            disabled={this.props.disabled}
-            onChange={this.handleGameTypeChange}
-          >
-            {allGameTypes ? (
-              allGameTypes.map(t => (
-                <MenuItem value={t.id} key={t.id}>
-                  {t.name}
-                </MenuItem>
-              ))
-            ) : (
-              <MenuItem disabled>Loading...</MenuItem>
-            )}
-          </TextField>
-        </FormControl>
-        <FormControl className={classes.formControl}>
-          <TextField
-            label="Name"
-            fullWidth
-            margin="normal"
-            variant="outlined"
-            value={name}
-            disabled={this.props.disabled}
-            onChange={this.handleNameChange}
-          />
-        </FormControl>
-        <FormControl className={classes.ratingFormControl}>
-          <Typography
-            component="legend"
-            color={this.props.disabled ? 'textSecondary' : 'initial'}
-          >
+      gameLoading ||
+      !allGameTypes!.find((t) => GameType.CompareObjects(t, type)) ||
+      maxDifficulty === 0 ||
+      maxHotLevel === 0 ||
+      name.length === 0 ||
+      isNaN(nbTurns) ||
+      nbTurns <= 0 ||
+      questionTypes.length === 0
+    );
+  };
+
+  const handleNameChange = (e: React.ChangeEvent<HTMLInputElement>) =>
+    props.editable && setName(e.target.value);
+
+  const handleTypeChange = (e: React.ChangeEvent<{ value: unknown }>) =>
+    props.editable &&
+    setType(
+      allGameTypes!.find((g) => g.id === parseInt(e.target.value as string))!
+    );
+
+  const handleMaxDifficultyChange = (
+    e: React.ChangeEvent<any>,
+    value: number | null
+  ) => props.editable && value != null && setMaxDifficulty(value);
+
+  const handleMaxHotLevelChange = (
+    e: React.ChangeEvent<any>,
+    value: number | null
+  ) => props.editable && value != null && setMaxHotLevel(value);
+
+  const handleNbTurnsChange = (e: React.ChangeEvent<HTMLInputElement>) =>
+    props.editable &&
+    // parseInt(e.target.value) >= 0 &&
+    // parseInt(e.target.value) < 1000000 &&
+    setNbTurns(parseInt(e.target.value));
+
+  const handleQuestionTypesChange = (value: QuestionType[]) =>
+    props.editable && setQuestionTypes(value);
+
+  const beforeSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+
+    if (!isDenied() && props.onSubmit) {
+      props.onSubmit(
+        Helper.clone(props.game, {
+          name,
+          type,
+          maxDifficulty,
+          maxHotLevel,
+          nbTurns,
+          questionTypes,
+        })
+      );
+    }
+  };
+
+  const beforeDelete = () => {
+    if (props.onDelete) props.onDelete(props.game.id);
+  };
+
+  const switchQuestionType = (questionTypeId: string) => {
+    const typeIndex = questionTypes.findIndex((qt) => qt.id === questionTypeId);
+
+    if (typeIndex !== -1) {
+      handleQuestionTypesChange(
+        questionTypes.filter((qt) => qt.id !== questionTypeId)
+      );
+    } else {
+      handleQuestionTypesChange([
+        ...questionTypes,
+        allQuestionsTypes.find((qt) => qt.id === questionTypeId)!,
+      ]);
+    }
+  };
+
+  return (
+    <form onSubmit={beforeSubmit}>
+      <TextField
+        style={{ textAlign: 'left', marginTop: 0 }}
+        name="typeSelect"
+        margin="normal"
+        variant="outlined"
+        select
+        required
+        disabled={props.disabled}
+        fullWidth
+        id="typeSelect"
+        label="Game type"
+        value={type.id}
+        onChange={handleTypeChange}
+      >
+        {allGameTypes ? (
+          allGameTypes.map((t) => (
+            <MenuItem value={t.id} key={t.id}>
+              {t.name}
+            </MenuItem>
+          ))
+        ) : (
+          <MenuItem disabled>Loading...</MenuItem>
+        )}
+      </TextField>
+      <TextField
+        name="name"
+        margin="normal"
+        variant="outlined"
+        type="text"
+        required
+        disabled={props.disabled}
+        fullWidth
+        id="name"
+        label="Name"
+        value={name}
+        onChange={handleNameChange}
+      />
+      <Grid container spacing={3} style={{ textAlign: 'center' }}>
+        <Grid item xs={6}>
+          <Typography component="legend" className={classes.unselectable}>
             Maximum&nbsp;Difficulty
           </Typography>
           <Rating
-            className={classes.ratingComponent}
-            name="difficulty"
+            disabled={props.disabled}
+            name="maxDifficulty"
             value={maxDifficulty}
-            onChange={this.handleMaxDifficultyChange}
-            disabled={this.props.disabled}
+            onChange={handleMaxDifficultyChange}
           />
-        </FormControl>
-        <FormControl className={classes.ratingFormControl}>
-          <Typography
-            component="legend"
-            color={this.props.disabled ? 'textSecondary' : 'initial'}
-          >
+        </Grid>
+        <Grid item xs={6}>
+          <Typography component="legend" className={classes.unselectable}>
             Maximum&nbsp;Hot&nbsp;Level
           </Typography>
           <Rating
-            name="hotLevel"
+            name="maxHotLevel"
+            disabled={props.disabled}
+            className={classes.hotLevelRating}
             value={maxHotLevel}
-            onChange={this.handleMaxHotLevelChange}
-            className={`${classes.ratingComponent} ${classes.hotLevelRating}`}
+            onChange={handleMaxHotLevelChange}
             icon={<FavoriteIcon fontSize="inherit" />}
-            disabled={this.props.disabled}
           />
-        </FormControl>
-        <FormControl className={classes.formControl}>
-          <TextField
-            label="Nb Turns"
-            fullWidth
-            margin="normal"
-            variant="outlined"
-            type="number"
-            disabled={this.props.disabled}
-            value={isNaN(nbTurns) ? '' : nbTurns}
-            onChange={this.handleNbTurnsChange}
-          />
-        </FormControl>
-        <FormControl className={classes.formControl}>
-          <InputLabel id="question-type-select">Question Types</InputLabel>
-          <Select
-            labelId="question-type-select"
-            multiple
-            input={<Input />}
-            value={questionTypes}
-            disabled={this.props.disabled}
-            onChange={this.handleQuestionTypeChange}
-            renderValue={selected => (
-              <div className={classes.chips}>
-                {(selected as QuestionType[]).map(value => (
-                  <Chip
-                    key={value.id}
-                    label={value.name}
-                    className={classes.chip}
-                    onDelete={() => this.removeQuestionType(value.id)}
-                    disabled={this.props.disabled}
-                  />
-                ))}
-              </div>
-            )}
-            MenuProps={{
-              transformOrigin: {
-                vertical: 'bottom',
-                horizontal: 'left'
-              },
-              PaperProps: {
-                style: {
-                  maxHeight: 300
-                }
+        </Grid>
+      </Grid>
+      <TextField
+        name="nbTurns"
+        margin="normal"
+        variant="outlined"
+        type="number"
+        required
+        disabled={props.disabled}
+        fullWidth
+        InputProps={{ inputProps: { min: 0, max: 999999 } }}
+        id="nbTurns"
+        label="Nb Turns"
+        value={nbTurns}
+        onChange={handleNbTurnsChange}
+      />
+      <OutlinedDiv label="Question Types" fullWidth>
+        {allQuestionsTypes
+          .filter((t) => t.hasQuestions)
+          .map((t) => (
+            <Chip
+              label={t.name}
+              key={t.id}
+              color={
+                questionTypes.find((type) =>
+                  QuestionType.CompareObjects(t, type)
+                )
+                  ? 'primary'
+                  : 'default'
               }
-            }}
-          >
-            {allQuestionTypes ? (
-              allQuestionTypes.map(questionType => (
-                <MenuItem key={questionType.id} value={questionType.id}>
-                  <Checkbox
-                    color="primary"
-                    checked={
-                      questionTypes.find(t =>
-                        QuestionType.CompareObjects(questionType, t)
-                      ) !== undefined
-                    }
-                  />
-                  <ListItemText primary={questionType.name} />
-                </MenuItem>
-              ))
-            ) : (
-              <MenuItem disabled>Loading...</MenuItem>
-            )}
-          </Select>
-        </FormControl>
-        <Button
-          type="submit"
-          variant="contained"
-          color="primary"
-          disabled={this.isDenied() || this.props.disabled}
-        >
-          {allGameTypes && allQuestionTypes ? this.props.buttonText : 'Loading'}
-        </Button>
-      </form>
-    );
-  }
-}
+              disabled={props.disabled}
+              className={classes.questionTypeChip}
+              onClick={() => switchQuestionType(t.id)}
+            />
+          ))}
+      </OutlinedDiv>
+      {(props.onDelete || props.onSubmit) && (
+        <Grid container spacing={2}>
+          {props.onDelete && (
+            <Grid item xs>
+              <DangerButton
+                fullWidth
+                variant="contained"
+                disabled={props.disabled}
+                onClick={beforeDelete}
+              >
+                {props.deleteButtonText}
+              </DangerButton>
+            </Grid>
+          )}
+          {props.onSubmit && (
+            <Grid item xs>
+              <Button
+                type="submit"
+                fullWidth
+                disabled={props.disabled || isDenied()}
+                variant="contained"
+                color="primary"
+              >
+                {props.acceptButtonText}
+              </Button>
+            </Grid>
+          )}
+        </Grid>
+      )}
+    </form>
+  );
+};
+
+GameFormComponent.defaultProps = {
+  deleteButtonText: 'Delete Button',
+  acceptButtonText: 'Accept Button',
+};
 
 const mapStateToProps = (states: RootState, ownProps: OwnProps): StateProps => {
   return {
     gameTypesState: states.gameTypesState,
     gamesState: states.gamesState,
-    questionTypesState: states.questionTypesState
+    questionTypesState: states.questionTypesState,
   };
 };
 
@@ -420,7 +315,7 @@ const mapDispatchToProps = (
   return {
     addError: async (error: any) => {
       await dispatch(addError(error));
-    }
+    },
   };
 };
 
@@ -429,4 +324,4 @@ export const GameForm = connect<StateProps, DispatchProps, OwnProps, RootState>(
   mapDispatchToProps,
   null,
   { forwardRef: true }
-)(withStyles(styles)(GameFormComponent));
+)(GameFormComponent);
