@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { ThunkDispatch } from 'redux-thunk';
 import {
@@ -14,20 +14,14 @@ import {
   IconButton,
   Button,
   Modal,
-  Tooltip
+  Tooltip,
 } from '@material-ui/core';
-import { withSnackbar, WithSnackbarProps } from 'notistack';
-import {
-  withStyles,
-  WithStyles,
-  createStyles,
-  StyleRules,
-  Theme
-} from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import DeleteIcon from '@material-ui/icons/Delete';
 import HourglassEmptyIcon from '@material-ui/icons/HourglassEmpty';
 import PersonIcon from '@material-ui/icons/Person';
+import SwipeableViews from 'react-swipeable-views';
 
 import { RootState } from '../../../store';
 import { addError } from '../../../store/errors/actions';
@@ -36,7 +30,7 @@ import { GameTypesState } from '../../../store/gameTypes/types';
 import { QuestionTypesState } from '../../../store/questionTypes/types';
 import { GamesActions } from '../../../store/games/actions';
 import { GameState } from '../../../store/game/types';
-import { Loading } from '../../components/loading/loading.component';
+import { Loading } from '../../components/utils/loading.component';
 import { User, GenderTable } from '../../../api/classes/user.class';
 import apiHandler from '../../../api/apiHandler';
 import { DummyUser } from '../../../api/classes/dummyUser.class';
@@ -45,25 +39,36 @@ import { UserState } from '../../../store/user/types';
 import { GamesState } from '../../../store/games/types';
 import { GameForm } from '../../components/game/game.form.component';
 import { DummyUserNew } from '../../components/user/dummy.new.component';
+import { LoadingOverlay } from '../../components/utils/loadingOverlay.component';
+import { GameActions } from '../../../store/game/actions';
 
-const styles = (theme: Theme): StyleRules =>
-  createStyles({
-    root: {
-      marginTop: theme.spacing(3)
-    },
-    addDummyUserButton: {
-      marginTop: theme.spacing(1),
-      marginBottom: theme.spacing(1)
-    }
-  });
+const useStyles = makeStyles((theme) => ({
+  root: {
+    marginTop: theme.spacing(3),
+  },
+  addDummyUserButton: {
+    marginTop: theme.spacing(1),
+    marginBottom: theme.spacing(1),
+  },
+  viewContainer: {
+    padding: theme.spacing(2),
+  },
+  selectUserContainer: {
+    height: 490,
+  },
+  tableContainer: {
+    paddingBottom: theme.spacing(1),
+  },
+}));
 
 interface OwnProps {
   displayId?: string;
 }
 
 interface DispatchProps {
-  gameUpdate: (game: Game) => Promise<any>;
+  gameUpdate: (game: Game, hideSuccess?: boolean) => Promise<any>;
   gameRemove: (gameId: string) => Promise<any>;
+  gameStartLoading: () => void;
   addError: (error: any) => void;
 }
 
@@ -75,228 +80,253 @@ interface StateProps {
   questionTypesState: QuestionTypesState;
 }
 
-type Props = StateProps &
-  OwnProps &
-  DispatchProps &
-  WithSnackbarProps &
-  WithStyles<typeof styles>;
+type Props = StateProps & OwnProps & DispatchProps;
 
-interface ComponentState {
-  selectedUser: User | null;
-  dummyModalOpen: boolean;
-  foundUsers: User[];
-}
+const GameCreatedPage: React.FunctionComponent<Props> = (props: Props) => {
+  const classes = useStyles();
 
-class GameCreatedPage extends React.Component<Props, ComponentState> {
-  /**
-   *
-   */
-  constructor(props: Props) {
-    super(props);
+  const [dummyModalOpen, setDummyModalOpen] = useState(false);
+  const [activeStep, setActiveStep] = useState(0);
+  const [selectedUser, setSelectedUser] = useState<User | undefined>(undefined);
+  const [foundUsers, setFoundUsers] = useState<User[]>([]);
+  const [game, setGame] = useState(
+    Helper.clone<Game>(props.gameState.game!, {})
+  );
+  const [gameStarting, setGameStarting] = useState(false);
 
-    this.state = {
-      selectedUser: null,
-      dummyModalOpen: false,
-      foundUsers: []
-    };
-  }
+  const isDisabled =
+    props.userState.user!.id !== props.gameState.game!.creator.id ||
+    props.gamesState.loading;
+  const cantStartGame = game.allUsers.length === 0 || isDisabled;
 
-  openDummyModal = () => {
-    this.setState({ dummyModalOpen: true });
-  };
+  useEffect(() => {
+    apiHandler.gameIo.joinGame(game.id);
 
-  closeDummyModal = () => {
-    this.setState({ dummyModalOpen: false });
-  };
-
-  handleSubmit = (game: Game) => {
-    this.props.gameUpdate(game);
-  };
-
-  addUser = async (user: User) => {
-    const { game } = this.props.gameState;
-
-    if (game) {
-      await this.props.gameUpdate(
-        Helper.clone(game, { users: [...game.users, user] })
-      );
-    }
-  };
-
-  removeUser = async (user: any) => {
-    const { game } = this.props.gameState;
-
-    if (game) {
-      await this.props.gameUpdate(
-        Helper.clone(game, {
-          users: game.users.filter(u => u.id !== user.id)
-        })
-      );
-    }
-  };
-
-  addDummyUser = async (dummy: DummyUser) => {
-    const { game } = this.props.gameState;
-
-    if (game) {
-      await this.props.gameUpdate(
-        Helper.clone(game, { dummyUsers: [...game.dummyUsers, dummy] })
-      );
-    }
-  };
-
-  removeDummyUser = async (dummy: any) => {
-    const { game } = this.props.gameState;
-
-    if (game) {
-      await this.props.gameUpdate(
-        Helper.clone(game, {
-          dummyUsers: game.dummyUsers.filter(d => d.id !== dummy.id)
-        })
-      );
-    }
-  };
-
-  handleSelectUser = (e: object, user: User) => {
-    this.setState({ selectedUser: null, foundUsers: [] }, () => {
-      (document.activeElement as any).blur();
+    apiHandler.gameIo.onceGameLoading((isLoading: boolean) => {
+      if (isLoading) {
+        props.gameStartLoading();
+      }
     });
 
-    this.addUser(user);
+    return () => {
+      apiHandler.gameIo.leaveGame(game.id);
+    };
+  }, []); // eslint-disable-line
+
+  const handleStepChange = (newStep: number) =>
+    newStep >= 0 && newStep <= steps.length && setActiveStep(newStep);
+
+  const handleNext = (game: Game) => {
+    setGame(game);
+
+    handleStepChange(activeStep + 1);
   };
 
-  handleSearchChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+  const handleBack = () => {
+    handleStepChange(activeStep - 1);
+  };
+
+  const startGame = async () => {
+    // TODO : do this server side
+
+    // const startedGame = Helper.clone(game, {
+    //   status: GameStatus.started,
+    //   scores: game.allUsers.map((u) => {
+    //     return Score.New({ score: 0, userId: u.id });
+    //   }),
+    // });
+
+    setGameStarting(true);
+
+    await props.gameUpdate(game, true);
+
+    apiHandler.gameIo.startGame(game.id);
+  };
+
+  const openDummyModal = () => setDummyModalOpen(true);
+  const closeDummyModal = () => setDummyModalOpen(false);
+
+  const handleSelectUser = (e: React.ChangeEvent<any>, user: User | null) => {
+    setFoundUsers([]);
+    setSelectedUser(undefined);
+
+    (document.activeElement as any).blur();
+
+    if (user != null) addUser(user);
+  };
+
+  const handleSearchChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     const searchName = e.target.value;
-    const { game } = this.props.gameState;
     apiHandler.userservice
       .findUsersByPartialName(searchName)
-      .then(users => {
-        this.setState({
-          foundUsers: game
-            ? users.filter(u => !game.users.map(gu => gu.id).includes(u.id))
+      .then((users) => {
+        setFoundUsers(
+          game
+            ? users.filter((u) => !game.users.map((gu) => gu.id).includes(u.id))
             : users
-        });
+        );
       })
-      .catch(e => {
-        this.setState({ foundUsers: [] });
+      .catch((e) => {
+        setFoundUsers([]);
       });
   };
 
-  handleDummyUserCreate = (dummy: DummyUser) => {
-    this.addDummyUser(dummy);
-
-    this.closeDummyModal();
+  const addUser = async (user: User) => {
+    setGame(Helper.clone(game, { users: [...game.users, user] }));
   };
 
-  render() {
-    const classes = this.props.classes;
-    const isDisabled =
-      this.props.userState.user!.id !== this.props.gameState.game!.creator.id ||
-      this.props.gamesState.loading;
+  const removeUser = async (user: any) => {
+    setGame(
+      Helper.clone(game, { users: game.users.filter((u) => u.id !== user.id) })
+    );
+  };
 
-    const game = this.props.gameState.game;
+  const addDummyUser = async (dummy: DummyUser) => {
+    setGame(Helper.clone(game, { dummyUsers: [...game.dummyUsers, dummy] }));
 
-    if (!game) return <Loading />;
+    closeDummyModal();
+  };
+
+  const removeDummyUser = async (dummy: any) => {
+    setGame(
+      Helper.clone(game, {
+        dummyUsers: game.dummyUsers.filter((d) => d.id !== dummy.id),
+      })
+    );
+  };
+
+  const steps = [
+    <GameForm
+      game={game}
+      onSubmit={handleNext}
+      editable
+      acceptButtonText="Next"
+      disabled={isDisabled}
+    />,
+    <div>
+      <Autocomplete
+        getOptionLabel={(option: User) => option.name}
+        options={foundUsers}
+        onChange={handleSelectUser}
+        value={selectedUser}
+        disabled={isDisabled}
+        renderInput={(params) => (
+          <TextField
+            {...params}
+            onChange={handleSearchChange}
+            label="Search an existing user"
+            variant="outlined"
+            fullWidth
+          />
+        )}
+      />
+      <Button
+        variant="contained"
+        color="primary"
+        fullWidth
+        className={classes.addDummyUserButton}
+        onClick={openDummyModal}
+        disabled={isDisabled}
+      >
+        Add a temporary user
+      </Button>
+
+      <TableContainer component="div" className={classes.tableContainer}>
+        <Table size="small" stickyHeader>
+          <TableHead>
+            <TableRow>
+              <TableCell>Type</TableCell>
+              <TableCell>User name</TableCell>
+              <TableCell align="right">Gender</TableCell>
+              <TableCell align="right">Actions</TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {game.allUsers.map((user) => {
+              const isDummy = user instanceof DummyUser; //isUUID.v1(user.id);
+              return (
+                <TableRow key={user.id}>
+                  <TableCell>
+                    {isDummy ? (
+                      <Tooltip title="Temporary user">
+                        <HourglassEmptyIcon />
+                      </Tooltip>
+                    ) : (
+                      <Tooltip title="Registered user">
+                        <PersonIcon />
+                      </Tooltip>
+                    )}
+                  </TableCell>
+                  <TableCell>{user.name}</TableCell>
+                  <TableCell align="right">
+                    {GenderTable[user.gender]}
+                  </TableCell>
+                  <TableCell align="right">
+                    <IconButton
+                      onClick={() =>
+                        isDummy ? removeDummyUser(user) : removeUser(user)
+                      }
+                      disabled={isDisabled}
+                    >
+                      <DeleteIcon />
+                    </IconButton>
+                  </TableCell>
+                </TableRow>
+              );
+            })}
+          </TableBody>
+        </Table>
+      </TableContainer>
+      <Grid container spacing={2}>
+        <Grid item xs={6}>
+          <Button
+            fullWidth
+            color="primary"
+            variant="contained"
+            onClick={handleBack}
+          >
+            Back
+          </Button>
+        </Grid>
+        <Grid item xs={6}>
+          <Button
+            fullWidth
+            color="primary"
+            variant="contained"
+            disabled={cantStartGame}
+            onClick={startGame}
+          >
+            Start game
+          </Button>
+        </Grid>
+      </Grid>
+    </div>,
+  ];
+
+  if (props.gameState.loading) {
+    return <Loading />;
+  } else {
     return (
       <>
+        <LoadingOverlay loading={gameStarting} />
         <Container maxWidth="md" className={classes.root}>
-          <Grid container spacing={1}>
-            <Grid item xs={12} md={6}>
-              <GameForm
-                game={game}
-                onSubmit={this.handleSubmit}
-                buttonText="Update Game"
-                disabled={isDisabled}
-              />
-            </Grid>
-            <Grid item xs={12} md={6}>
-              <Autocomplete
-                getOptionLabel={(option: User) => option.name}
-                options={this.state.foundUsers}
-                onChange={this.handleSelectUser}
-                value={this.state.selectedUser}
-                disabled={isDisabled}
-                renderInput={params => (
-                  <TextField
-                    {...params}
-                    onChange={this.handleSearchChange}
-                    label="Search an existing user"
-                    variant="outlined"
-                    fullWidth
-                  />
-                )}
-              />
-              <Button
-                variant="contained"
-                color="primary"
-                fullWidth
-                className={classes.addDummyUserButton}
-                onClick={this.openDummyModal}
-                disabled={isDisabled}
-              >
-                Add a temporary user
-              </Button>
-
-              <TableContainer component="div">
-                <Table size="small" stickyHeader>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Type</TableCell>
-                      <TableCell>User name</TableCell>
-                      <TableCell align="right">Gender</TableCell>
-                      <TableCell align="right">Actions</TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {game.allUsers.map(user => {
-                      const isDummy = user instanceof DummyUser; //isUUID.v1(user.id);
-                      return (
-                        <TableRow key={user.id}>
-                          <TableCell>
-                            {isDummy ? (
-                              <Tooltip title="Temporary user">
-                                <HourglassEmptyIcon />
-                              </Tooltip>
-                            ) : (
-                              <Tooltip title="Registered user">
-                                <PersonIcon />
-                              </Tooltip>
-                            )}
-                          </TableCell>
-                          <TableCell>{user.name}</TableCell>
-                          <TableCell align="right">
-                            {GenderTable[user.gender]}
-                          </TableCell>
-                          <TableCell align="right">
-                            <IconButton
-                              color="primary"
-                              onClick={() =>
-                                isDummy
-                                  ? this.removeDummyUser(user)
-                                  : this.removeUser(user)
-                              }
-                              disabled={isDisabled}
-                            >
-                              <DeleteIcon />
-                            </IconButton>
-                          </TableCell>
-                        </TableRow>
-                      );
-                    })}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-              {this.props.gamesState.loading && <Loading />}
-            </Grid>
-          </Grid>
+          <SwipeableViews index={activeStep} disabled>
+            {steps.map((step, index) => {
+              return (
+                <div className={classes.viewContainer} key={index}>
+                  {step}
+                </div>
+              );
+            })}
+          </SwipeableViews>
         </Container>
-        <Modal open={this.state.dummyModalOpen} onClose={this.closeDummyModal}>
-          <DummyUserNew dummyUserCreate={this.handleDummyUserCreate} />
+        <Modal open={dummyModalOpen} onClose={closeDummyModal}>
+          <DummyUserNew dummyUserCreate={addDummyUser} />
         </Modal>
       </>
     );
   }
-}
+};
 
 const mapStateToProps = (states: RootState, ownProps: OwnProps): StateProps => {
   return {
@@ -304,7 +334,7 @@ const mapStateToProps = (states: RootState, ownProps: OwnProps): StateProps => {
     gamesState: states.gamesState,
     gameState: states.gameState,
     questionTypesState: states.questionTypesState,
-    userState: states.userState
+    userState: states.userState,
   };
 };
 
@@ -313,15 +343,18 @@ const mapDispatchToProps = (
   ownProps: OwnProps
 ): DispatchProps => {
   return {
-    gameUpdate: async (game: Game) => {
-      return await dispatch(GamesActions.gameUpdate(game));
+    gameUpdate: async (game: Game, hideSuccess?: boolean) => {
+      return await dispatch(GamesActions.gameUpdate(game, hideSuccess));
     },
     gameRemove: async (gameId: string) => {
       return await dispatch(GamesActions.gameRemove(gameId));
     },
+    gameStartLoading: () => {
+      dispatch(GameActions.gameStartLoading());
+    },
     addError: async (error: any) => {
       await dispatch(addError(error));
-    }
+    },
   };
 };
 
@@ -330,6 +363,7 @@ export const GameCreated = connect<
   DispatchProps,
   OwnProps,
   RootState
->(mapStateToProps, mapDispatchToProps, null, { forwardRef: true })(
-  withStyles(styles)(withSnackbar(GameCreatedPage))
-);
+>(
+  mapStateToProps,
+  mapDispatchToProps
+)(GameCreatedPage);
