@@ -2,6 +2,9 @@ import { QuestionType, QuestionTypeBackModel } from './questionType.class';
 import { User, UserBackModel } from './user.class';
 import { GameType, GameTypeBackModel } from './gameType.class';
 import { DummyUser } from './dummyUser.class';
+import { Helper } from '../../helper';
+import { Question, QuestionBackModel } from './question.class';
+import { Score } from './score.class';
 
 export interface GameBackModel {
   _id: string;
@@ -15,18 +18,20 @@ export interface GameBackModel {
   _maxHotLevel: number;
   _creator: UserBackModel;
   _type: GameTypeBackModel;
+  _scores: Score[];
+  _actualQuestion: QuestionBackModel;
 }
 
 export enum GameStatus {
   created = 'Created',
   started = 'Started',
-  finished = 'Finished'
+  finished = 'Finished',
 }
 
 export enum GameStatusColors {
   Created = 'default',
   Started = 'primary',
-  Finished = 'secondary'
+  Finished = 'secondary',
 }
 
 export class Game {
@@ -56,6 +61,10 @@ export class Game {
 
   public status: GameStatus = GameStatus.created;
 
+  public scores: Score[] = [];
+
+  public actualQuestion: Question = new Question();
+
   public get allUsers() {
     return [...this.users, ...this.dummyUsers];
   }
@@ -63,6 +72,26 @@ export class Game {
   public canStart(): boolean {
     if (this.allUsers.length <= 1) return false;
     return true;
+  }
+
+  public getActualPlayerIndex(): number {
+    return Helper.getPlayer(
+      this.actualTurn,
+      this.nbTurns,
+      this.allUsers.length,
+      this.displayId
+    );
+  }
+
+  public getActualplayer(): User | DummyUser {
+    let userIndex = Helper.getPlayer(
+      this.actualTurn,
+      this.nbTurns,
+      this.allUsers.length,
+      this.displayId
+    );
+
+    return this.allUsers[userIndex];
   }
 
   public static New(datas: Partial<Game>): Game {
@@ -92,6 +121,10 @@ export class Game {
     newObj.creator = User.fromBack(datas._creator);
     newObj.type = GameType.fromBack(datas._type);
     newObj.status = datas._status;
+    for (const scoreModel of datas._scores) {
+      newObj.scores.push(Score.fromBack(scoreModel));
+    }
+    newObj.actualQuestion = Question.fromBack(datas._actualQuestion);
 
     return newObj;
   }
@@ -118,7 +151,9 @@ export class Game {
       obj1.maxHotLevel === obj2.maxHotLevel &&
       User.CompareObjects(obj1.creator, obj2.creator) &&
       GameType.CompareObjects(obj1.type, obj2.type) &&
-      obj1.status === obj2.status
+      obj1.status === obj2.status &&
+      Score.CompareArrays(obj1.scores, obj2.scores) &&
+      Question.CompareObjects(obj1.actualQuestion, obj2.actualQuestion)
     );
   }
 }
