@@ -1,6 +1,15 @@
+import {
+  QuestionTemplate,
+  QuestionTemplateBackModel,
+} from './questionTemplate.class';
+
 export interface QuestionTypeBackModel {
   _id: string;
   _name: string;
+  _description: string;
+  _template: QuestionTemplateBackModel;
+  _icon: string;
+  _hasQuestions: boolean;
 }
 
 export class QuestionType {
@@ -8,9 +17,13 @@ export class QuestionType {
 
   public name: string = '';
 
-  public temp(): string {
-    return 'ok';
-  }
+  public description: string = '';
+
+  public template: QuestionTemplate = new QuestionTemplate();
+
+  public icon: string = '';
+
+  public hasQuestions: boolean = false;
 
   public static New(datas: Partial<QuestionType>): QuestionType {
     return Object.assign(new QuestionType(), datas);
@@ -21,6 +34,10 @@ export class QuestionType {
 
     newObj.id = datas._id;
     newObj.name = datas._name;
+    newObj.description = datas._description;
+    newObj.template = QuestionTemplate.fromBack(datas._template);
+    newObj.icon = datas._icon;
+    newObj.hasQuestions = datas._hasQuestions;
 
     return newObj;
   }
@@ -43,7 +60,11 @@ export class QuestionType {
       obj1 !== undefined &&
       obj2 !== undefined &&
       obj1.id === obj2.id &&
-      obj1.name === obj2.name
+      obj1.name === obj2.name &&
+      obj1.description === obj2.description &&
+      QuestionTemplate.CompareObjects(obj1.template, obj2.template) &&
+      obj1.icon === obj2.icon &&
+      obj1.hasQuestions === obj2.hasQuestions
     );
   }
 }
