@@ -1,12 +1,11 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import {
   Dialog,
   DialogTitle,
   DialogContent,
   DialogContentText,
   DialogActions,
-  Button
+  Button,
 } from '@material-ui/core';
 import { EventEmitter } from 'events';
 //import { makeStyles } from '@material-ui/core';
@@ -35,7 +34,7 @@ class YesNoComponent extends React.Component<Props, ComponentState> {
     title: 'Are you sure ?',
     text: '',
     acceptText: 'Accept',
-    denyText: 'Deny'
+    denyText: 'Deny',
   };
 
   /**
@@ -66,7 +65,7 @@ class YesNoComponent extends React.Component<Props, ComponentState> {
       title: newState.title!,
       text: newState.text!,
       acceptText: newState.acceptText!,
-      denyText: newState.denyText!
+      denyText: newState.denyText!,
     });
     return new Promise((res, rej) => {
       this.events.once('resolve', () => {
@@ -103,10 +102,19 @@ class YesNoComponent extends React.Component<Props, ComponentState> {
           </DialogContent>
         )}
         <DialogActions>
-          <Button onClick={this.handleAccept} color="primary">
+          <Button
+            onClick={this.handleAccept}
+            color="primary"
+            variant="contained"
+          >
             {this.state.acceptText}
           </Button>
-          <Button onClick={this.handleDeny} color="primary" autoFocus>
+          <Button
+            onClick={this.handleDeny}
+            autoFocus
+            color="primary"
+            variant="contained"
+          >
             {this.state.denyText}
           </Button>
         </DialogActions>
@@ -114,11 +122,8 @@ class YesNoComponent extends React.Component<Props, ComponentState> {
     );
   }
 }
-const ref = React.createRef<YesNoComponent>();
+export const yesNoRef = React.createRef<YesNoComponent>();
 
-ReactDOM.render(
-  <YesNoComponent ref={ref} />,
-  document.getElementById('yesnoRoot')
-);
+export const YesNo = YesNoComponent;
 
-export const yesNoController = ref.current;
+export const yesNoController = () => yesNoRef.current;
