@@ -1,9 +1,9 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { ThunkDispatch } from 'redux-thunk';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import { RootState } from '../../../store';
-import { Loading } from '../../components/loading/loading.component';
+import { Loading } from '../utils/loading.component';
 import { GameState } from '../../../store/game/types';
 import { GameActions } from '../../../store/game/actions';
 import { GameStatus } from '../../../api/classes/game.class';
@@ -11,6 +11,7 @@ import { GameStatus } from '../../../api/classes/game.class';
 import { UserState } from '../../../store/user/types';
 import { GameNotFound } from '../../pages/games/game.notFound.page';
 import { GameCreated } from '../../pages/games/game.created.page';
+import { GameStarted } from '../../pages/games/game.started.page';
 
 interface RouteParams {
   displayId: string;
@@ -31,54 +32,43 @@ type Props = StateProps & OwnProps & DispatchProps & RouteComponentProps;
 
 interface ComponentState {}
 
-class GameMiddlewareComponent extends React.Component<Props, ComponentState> {
-  /**
-   *
-   */
-  constructor(props: Props) {
-    super(props);
+const GameMiddlewareComponent: React.FunctionComponent<Props> = (
+  props: Props
+) => {
+  const { displayId } = props.match.params as any;
 
-    this.state = {};
+  const { game, loading: gameLoading } = props.gameState;
+  const { user, loading: userLoading } = props.userState;
 
-    const { displayId } = this.props.match.params as any;
-    const { game, loading } = this.props.gameState;
+  useEffect(() => {
+    if (displayId) props.gameGetByDisplayId(displayId);
+  }, []); // eslint-disable-line
 
-    if (!game && !loading && displayId) {
-      this.props.gameGetByDisplayId(displayId);
-    }
-  }
-
-  render() {
-    const { displayId } = this.props.match.params as any;
-    const { game, loading: gameLoading } = this.props.gameState;
-    const { user, loading: userLoading } = this.props.userState;
-
-    if (game && user) {
-      if (
-        game.creator.id === user.id ||
-        game.users.map(u => u.id).includes(user.id)
-      ) {
-        switch (game.status) {
-          case GameStatus.created:
-            return <GameCreated displayId={displayId} />;
-          case GameStatus.started:
-            return <>started</>;
-          case GameStatus.finished:
-            return <>finished</>;
-        }
-      } else {
-        return <GameNotFound displayId={displayId} />;
+  if (game && user) {
+    if (
+      game.creator.id === user.id ||
+      game.users.map((u) => u.id).includes(user.id)
+    ) {
+      switch (game.status) {
+        case GameStatus.created:
+          return <GameCreated displayId={displayId} />;
+        case GameStatus.started:
+          return <GameStarted displayId={displayId} />;
+        case GameStatus.finished:
+          return <>finished</>;
       }
-    } else if (gameLoading || userLoading) return <Loading />;
-    else if (!game) return <GameNotFound displayId={displayId} />;
-    return <>An error occured</>;
-  }
-}
+    } else {
+      return <GameNotFound displayId={displayId} />;
+    }
+  } else if (gameLoading || userLoading) return <Loading />;
+  else if (!game) return <GameNotFound displayId={displayId} />;
+  return <>An error occured</>;
+};
 
 const mapStateToProps = (states: RootState, ownProps: OwnProps): StateProps => {
   return {
     gameState: states.gameState,
-    userState: states.userState
+    userState: states.userState,
   };
 };
 
@@ -89,7 +79,7 @@ const mapDispatchToProps = (
   return {
     gameGetByDisplayId: async (displayId: string) => {
       dispatch(GameActions.gameGetByDisplayId(displayId));
-    }
+    },
   };
 };
 
