@@ -6,8 +6,7 @@ import { RootState } from '../../../store';
 import { UserState } from '../../../store/user/types';
 import { relog } from '../../../store/user/actions';
 //import { makeStyles } from '@material-ui/core';
-import { Loading } from '../loading/loading.component';
-import { render } from 'react-dom';
+import { Loading } from '../utils/loading.component';
 
 /*
 const useStyles = makeStyles(theme => ({
@@ -48,7 +47,7 @@ type Props = StateProps & OwnProps & DispatchProps;
 const GuardComponent: React.FunctionComponent<Props> = (props: Props) => {
   const idViceAllowed =
     props.idViceAllowed === undefined ? true : props.idViceAllowed;
-  const { minimalPermission, children, redirect } = props;
+  const { minimalPermission } = props;
   const { user, loading } = props.userState;
   //const classes = useStyles();
 
