@@ -2,11 +2,11 @@ import React from 'react';
 import { Container, Typography, makeStyles } from '@material-ui/core';
 import { withRouter, RouteComponentProps } from 'react-router';
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles((theme) => ({
   root: {},
   text: {
-    marginTop: theme.spacing(5)
-  }
+    marginTop: theme.spacing(5),
+  },
 }));
 
 interface OwnProps {
@@ -18,7 +18,7 @@ type Props = OwnProps & RouteComponentProps;
 const GameNotFoundPage: React.FunctionComponent<Props> = (props: Props) => {
   const classes = useStyles();
 
-  const displayId = props.displayId || props.location.state.displayId;
+  const displayId = props.displayId || (props.location.state! as any).displayId;
 
   return (
     <Container maxWidth="md" className={classes.root}>
