@@ -11,7 +11,9 @@ export enum GameActionTypes {
   ACTION_FAILURE = '@@game/ACTION_FAILURE',
   GET = '@@game/GET',
   UPDATE = '@@game/UPDATE',
-  REMOVE = '@@game/REMOVE'
+  REMOVE = '@@game/REMOVE',
+  START_LOADING = '@@game/START_LOADING',
+  FINISH_LOADING = '@@game/FINISH_LOADING',
 }
 
 // Declare state types with `readonly` modifier to get compile time immutability.
